# v1/api/static/swagger.json
{
    "openapi": "3.0.0",
  "info": {
    "title": "Event Invitation Management Application API - OpenAPI 3.0",
    "description": "This application is designed to manage the event invitations for event owners and planners.",
    "contact": {
      "email": "justinoghenekomebedi@gmail.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/v1/api"
    }
  ],
...
  "paths": {
...
    "/events/": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create an event",
        "operationId": "8",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_name": {
                    "type": "string"
                  },
                  "event_owner": {
                    "type": "string"
                  },
                  "event_location": {
                    "type": "string"
                  },
                  "date_time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                        "type": "object",
                        "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/{event_id}": {
      "put": {
        "tags": [
          "events"
        ],
        "summary": "Update an event",
        "operationId": "9",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_name": {
                    "type": "string"
                  },
                  "event_owner": {
                    "type": "string"
                  },
                  "event_location": {
                    "type": "string"
                  },
                  "date_time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                        "type": "object",
                        "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/delete_event/{event_id}": {
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete an event",
        "operationId": "10",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "result": {
                        "type": "object",
                        "$ref": "#/components/schemas/Event"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing Authorization Header"
          },
          "404": {
            "description": "Event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Event": {
        "type": "object",
        "properties": {
          "event_name": {
            "type": "string",
            "example": "Product launch"
          },
          "event_location": {
            "type": "string",
            "example": "Transcorp Hilton, Abuja"
          },
          "date_time": {
            "type": "string",
            "example": "24-09-17T:09:30:00.0000"
          },
          "invitation_link": {
            "type": "string",
            "example": "invitations/rsvp/10/100"
          }
        },
        "xml": {
          "name": "event"
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}